refactor: フォルダ構成を整理。Productに関するフォルダにプロダクトを選択するダイアログを移動。
