+++ src/components/product/ProductSelectorDialog.vue
